Initial implementation
Set up the base project. Have a feeder framework that manages the Python
snippets to feed the dashboard data and add sample jobs for the default
Smashing dashboard layout. Add the default Smashing dashboard layout to
dashboard/dashboards and add an empty dashboard/widgets directory to
allow for custom widgets to be added to the dashboard.

Update README with some basic instructions.